docs: update based on PR feedback
merges existing drag-drop section
