+++ packages/website/docs/02_components/display/drag_and_drop/overview.mdx
@@ -128,14 +128,20 @@ the visual changes with drop-to-remove interactions.
 
 ## Portalled items
 
-When using EUI drag components inside a stacking context, the fixed drag positioning might be affected. To ensure 
-dragging works as expected inside e.g. **EuiFlyout** or **EuiModal use the prop `usePortal` on **EuiDraggable** components. 
-This will render the currently dragged element inside a portal on the body scope.
+**EuiDraggables** use fixed positioning to render and animate the item being dragged.
+This positioning logic does not work as expected when used inside of containers that have their own
+[stacking context](https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_positioned_layout/Understanding_z-index/Stacking_context).
 
+To ensure dragging works as expected inside e.g. **EuiFlyout** or **EuiModal use the prop `usePortal` on **EuiDraggable** components. 
+This will render the currently dragged element inside a portal appended to the document body (or wherever
+**EuiPortal** is configured to `insert` to by default).
+
+:::warning 
 If the styling of the **EuiDraggable** content is relative to an outer scope component, the styling won't be applied
 whiled dragging it when using **usePortal**. This is because due to the changed position in the DOM which changes
 previous hierarchical relations to other ancestor elements. To prevent this from happening, we recommend to apply 
 styling from within the **EuiDraggable** scope.
+:::
 
 ```tsx interactive
 import React, { useState } from 'react';
@@ -171,6 +177,8 @@ const makeList = (number, start = 1) =>
 export default () => {
   const [isFlyoutOpen, setFlyoutOpen] = useState(false);
   const [isModalOpen, setModalOpen] = useState(false);
+  const [isPopoverOpen, setIsPopoverOpen] = useState(false);
+
   const [list, setList] = useState(makeList(3));
   const onDragEnd = ({ source, destination }) => {
     if (source && destination) {
@@ -256,6 +264,49 @@ export default () => {
           </EuiModalBody>
         </EuiModal>
       )}
+
+      <EuiSpacer />
+
+      <EuiPopover
+        isOpen={isPopoverOpen}
+        closePopover={() => setIsPopoverOpen(false)}
+        button={
+          <EuiButton onClick={() => setIsPopoverOpen(!isPopoverOpen)}>
+            Toggle popover
+          </EuiButton>
+        }
+        panelPaddingSize="none"
+        panelProps={{ css: { inlineSize: 200 } }}
+      >
+        <EuiDragDropContext
+          onDragEnd={({ source, destination }) => {
+            if (source && destination) {
+              const items = euiDragDropReorder(
+                list,
+                source.index,
+                destination.index
+              );
+              setList(items);
+            }
+          }}
+        >
+          <EuiDroppable droppableId="droppableInPopover" spacing="m">
+            {list.map(({ content, id }, idx) => (
+              <EuiDraggable
+                spacing="m"
+                key={id}
+                index={idx}
+                draggableId={id}
+                usePortal
+              >
+                {(provided, state) => (
+                  <EuiPanel hasShadow={state.isDragging}>{content}</EuiPanel>
+                )}
+              </EuiDraggable>
+            ))}
+          </EuiDroppable>
+        </EuiDragDropContext>
+      </EuiPopover>
     </>
   );
 };
@@ -267,14 +318,3 @@ export default () => {
 movement, vertical movement, flexbox, panel inception, you name it.
 
 <!-- TODO: Add kitchen sink example -->
-
-## Using drag and drop in popovers
-
-**EuiDraggables** use fixed positioning to render and animate the item being dragged.
-This positioning logic does not work as expected when used inside of containers that have their own
-[stacking context](https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_positioned_layout/Understanding_z-index/Stacking_context).
-
-This behavior particularly affects [**EuiPopover**](#/layout/popover). If using drag and drop UX within a popover,
-you **must** include the `<EuiPopover hasDragDrop>` prop for items to properly render while being dragged.
-
-<!-- TODO: Add drag and drop in popover example -->
